+++ custom_components/spotifyplus/intent_handlers/spotifyplusplayermediapause_handler.py
@@ -11,7 +11,6 @@ from homeassistant.helpers.intent import (
     _SlotsType,
 )
 from homeassistant.const import (
-    STATE_PAUSED,
     STATE_PLAYING,
 )
 
@@ -25,18 +24,18 @@ from ..const import (
     CONF_VALUE,
     DOMAIN,
     PLATFORM_SPOTIFYPLUS,
-    INTENT_PLAYER_MEDIA_SKIP_NEXT,
+    INTENT_PLAYER_MEDIA_PAUSE,
     RESPONSE_ERROR_UNHANDLED,
     RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
-    SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT,
+    SERVICE_SPOTIFY_PLAYER_MEDIA_PAUSE,
 )
 
 from .spotifyplusintenthandler import SpotifyPlusIntentHandler
 
 
-class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
+class SpotifyPlusPlayerMediaPause_Handler(SpotifyPlusIntentHandler):
     """
-    Handles intents for SpotifyPlusPlayerMediaSkipNext.
+    Handles intents for SpotifyPlusPlayerMediaPause.
     """
     def __init__(self) -> None:
         """
@@ -46,8 +45,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
         super().__init__()
 
         # set intent handler basics.
-        self.description = "Skips to next track in the user's queue for the specified SpotifyPlus media player."
-        self.intent_type = INTENT_PLAYER_MEDIA_SKIP_NEXT
+        self.description = "Pause media play for the specified SpotifyPlus media player."
+        self.intent_type = INTENT_PLAYER_MEDIA_PAUSE
         self.platforms = {PLATFORM_SPOTIFYPLUS}
 
 
@@ -98,8 +97,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
                 intentObj,
                 intentResponse,
                 slots=intentObj.slots,
-                desiredFeatures=MediaPlayerEntityFeature.NEXT_TRACK | MediaPlayerEntityFeature.PLAY_MEDIA,
-                desiredStates=[STATE_PLAYING, STATE_PAUSED],
+                desiredFeatures=MediaPlayerEntityFeature.PAUSE | MediaPlayerEntityFeature.PLAY_MEDIA,
+                desiredStates=[STATE_PLAYING],
                 desiredStateResponseKey=RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
             )
 
@@ -112,7 +111,7 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
             delay = slots.get("delay", {}).get(CONF_VALUE, None)
 
             # set service name and build parameters.
-            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT
+            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_PAUSE
             svcData:dict = \
             {
                 "entity_id": playerEntityState.entity_id,