+++ custom_components/spotifyplus/intent_handlers/spotifyplusplayermediaresume_handler.py
@@ -11,8 +11,11 @@ from homeassistant.helpers.intent import (
     _SlotsType,
 )
 from homeassistant.const import (
+    STATE_ON,
+    STATE_IDLE,
+    STATE_OFF,
     STATE_PAUSED,
-    STATE_PLAYING,
+    STATE_UNKNOWN,
 )
 
 from smartinspectpython.siauto import SILevel, SIColors
@@ -25,18 +28,18 @@ from ..const import (
     CONF_VALUE,
     DOMAIN,
     PLATFORM_SPOTIFYPLUS,
-    INTENT_PLAYER_MEDIA_SKIP_NEXT,
+    INTENT_PLAYER_MEDIA_RESUME,
     RESPONSE_ERROR_UNHANDLED,
     RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
-    SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT,
+    SERVICE_SPOTIFY_PLAYER_MEDIA_RESUME,
 )
 
 from .spotifyplusintenthandler import SpotifyPlusIntentHandler
 
 
-class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
+class SpotifyPlusPlayerMediaResume_Handler(SpotifyPlusIntentHandler):
     """
-    Handles intents for SpotifyPlusPlayerMediaSkipNext.
+    Handles intents for SpotifyPlusPlayerMediaResume.
     """
     def __init__(self) -> None:
         """
@@ -46,8 +49,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
         super().__init__()
 
         # set intent handler basics.
-        self.description = "Skips to next track in the user's queue for the specified SpotifyPlus media player."
-        self.intent_type = INTENT_PLAYER_MEDIA_SKIP_NEXT
+        self.description = "Resume media play for the specified SpotifyPlus media player."
+        self.intent_type = INTENT_PLAYER_MEDIA_RESUME
         self.platforms = {PLATFORM_SPOTIFYPLUS}
 
 
@@ -98,8 +101,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
                 intentObj,
                 intentResponse,
                 slots=intentObj.slots,
-                desiredFeatures=MediaPlayerEntityFeature.NEXT_TRACK | MediaPlayerEntityFeature.PLAY_MEDIA,
-                desiredStates=[STATE_PLAYING, STATE_PAUSED],
+                desiredFeatures=MediaPlayerEntityFeature.PAUSE | MediaPlayerEntityFeature.PLAY_MEDIA,
+                desiredStates=[STATE_PAUSED, STATE_OFF, STATE_ON, STATE_IDLE, STATE_UNKNOWN],
                 desiredStateResponseKey=RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
             )
 
@@ -112,7 +115,7 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
             delay = slots.get("delay", {}).get(CONF_VALUE, None)
 
             # set service name and build parameters.
-            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT
+            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_RESUME
             svcData:dict = \
             {
                 "entity_id": playerEntityState.entity_id,