+++ custom_components/spotifyplus/intent_handlers/spotifyplusplayermediaskipstart_handler.py
@@ -25,18 +25,18 @@ from ..const import (
     CONF_VALUE,
     DOMAIN,
     PLATFORM_SPOTIFYPLUS,
-    INTENT_PLAYER_MEDIA_SKIP_NEXT,
+    INTENT_PLAYER_MEDIA_SKIP_START,
     RESPONSE_ERROR_UNHANDLED,
     RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
-    SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT,
+    SERVICE_SPOTIFY_PLAYER_MEDIA_SEEK,
 )
 
 from .spotifyplusintenthandler import SpotifyPlusIntentHandler
 
 
-class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
+class SpotifyPlusPlayerMediaSkipStart_Handler(SpotifyPlusIntentHandler):
     """
-    Handles intents for SpotifyPlusPlayerMediaSkipNext.
+    Handles intents for SpotifyPlusPlayerMediaSkipStart.
     """
     def __init__(self) -> None:
         """
@@ -46,8 +46,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
         super().__init__()
 
         # set intent handler basics.
-        self.description = "Skips to next track in the user's queue for the specified SpotifyPlus media player."
-        self.intent_type = INTENT_PLAYER_MEDIA_SKIP_NEXT
+        self.description = "Restarts the currently playing track for the specified SpotifyPlus media player."
+        self.intent_type = INTENT_PLAYER_MEDIA_SKIP_START
         self.platforms = {PLATFORM_SPOTIFYPLUS}
 
 
@@ -98,7 +98,7 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
                 intentObj,
                 intentResponse,
                 slots=intentObj.slots,
-                desiredFeatures=MediaPlayerEntityFeature.NEXT_TRACK | MediaPlayerEntityFeature.PLAY_MEDIA,
+                desiredFeatures=MediaPlayerEntityFeature.SEEK | MediaPlayerEntityFeature.PLAY_MEDIA,
                 desiredStates=[STATE_PLAYING, STATE_PAUSED],
                 desiredStateResponseKey=RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
             )
@@ -112,11 +112,12 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
             delay = slots.get("delay", {}).get(CONF_VALUE, None)
 
             # set service name and build parameters.
-            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT
+            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_SEEK
             svcData:dict = \
             {
                 "entity_id": playerEntityState.entity_id,
                 "device_id": "",  # always use current device for this service call.
+                "position_ms": 0, # restart track
                 "delay": delay
             }