+++ custom_components/spotifyplus/intent_handlers/spotifyplusvolumesetstep_handler.py
@@ -25,18 +25,18 @@ from ..const import (
     CONF_VALUE,
     DOMAIN,
     PLATFORM_SPOTIFYPLUS,
-    INTENT_PLAYER_MEDIA_SKIP_NEXT,
+    INTENT_VOLUME_SET_STEP,
     RESPONSE_ERROR_UNHANDLED,
     RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
-    SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT,
+    SERVICE_VOLUME_SET_STEP,
 )
 
 from .spotifyplusintenthandler import SpotifyPlusIntentHandler
 
 
-class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
+class SpotifyPlusVolumeSetStep_Handler(SpotifyPlusIntentHandler):
     """
-    Handles intents for SpotifyPlusPlayerMediaSkipNext.
+    Handles intents for SpotifyPlusVolumeSetStep.
     """
     def __init__(self) -> None:
         """
@@ -46,8 +46,8 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
         super().__init__()
 
         # set intent handler basics.
-        self.description = "Skips to next track in the user's queue for the specified SpotifyPlus media player."
-        self.intent_type = INTENT_PLAYER_MEDIA_SKIP_NEXT
+        self.description = "Set level used for volume step services for the specified SpotifyPlus media player."
+        self.intent_type = INTENT_VOLUME_SET_STEP
         self.platforms = {PLATFORM_SPOTIFYPLUS}
 
 
@@ -66,7 +66,7 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
             vol.Optional("preferred_floor_id"): cv.string,
 
             # slots for other service arguments.
-            vol.Optional("delay", default=0.50): vol.Any(None, vol.All(vol.Coerce(float), vol.Range(min=0, max=10.0)))
+            vol.Optional("player_volume_step_pct", default=10): vol.Any(None, vol.All(vol.Coerce(int), vol.Range(min=1, max=100))),
         }
 
 
@@ -98,9 +98,9 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
                 intentObj,
                 intentResponse,
                 slots=intentObj.slots,
-                desiredFeatures=MediaPlayerEntityFeature.NEXT_TRACK | MediaPlayerEntityFeature.PLAY_MEDIA,
-                desiredStates=[STATE_PLAYING, STATE_PAUSED],
-                desiredStateResponseKey=RESPONSE_PLAYER_NOT_PLAYING_MEDIA,
+                desiredFeatures=MediaPlayerEntityFeature.VOLUME_STEP,
+                desiredStates=None,
+                desiredStateResponseKey=None,
             )
 
             # if media player was not resolved, then we are done;
@@ -109,15 +109,14 @@ class SpotifyPlusPlayerMediaSkipNext_Handler(SpotifyPlusIntentHandler):
                 return intentResponse
             
             # get optional arguments (if provided).
-            delay = slots.get("delay", {}).get(CONF_VALUE, None)
+            player_volume_step_pct = slots.get("player_volume_step_pct", {}).get(CONF_VALUE, 10)
 
             # set service name and build parameters.
-            svcName:str = SERVICE_SPOTIFY_PLAYER_MEDIA_SKIP_NEXT
+            svcName:str = SERVICE_VOLUME_SET_STEP
             svcData:dict = \
             {
                 "entity_id": playerEntityState.entity_id,
-                "device_id": "",  # always use current device for this service call.
-                "delay": delay
+                "level_percent": player_volume_step_pct,
             }
 
             # call integration service for this intent.